% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btf_coord.R
\name{btf_coord}
\alias{btf_coord}
\title{Coordinate decomposition}
\usage{
btf_coord(x, ...)
}
\description{
Collect all coordinates in one table, the path_link_vertex table
has the information about the original grouping.
}
