% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btf_path.R
\name{btf_PATH}
\alias{btf_PATH}
\title{Basic Topological PATH Form}
\usage{
btf_PATH(x, ...)
}
\arguments{
\item{...}{}
}
\description{
The PATH form is a simple normalization by vertex, it builds a common vertex pool
and provides the link tables (in order) to paths and objects.
}
