% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibble-package.R, R/gibble.R
\docType{package}
\name{gibble}
\alias{gibble}
\alias{gibble-package}
\alias{gibble}
\alias{gibble}
\alias{gibble.default}
\alias{gibble}
\alias{gibble.POINT}
\alias{gibble}
\alias{gibble.MULTIPOINT}
\alias{gibble}
\alias{gibble.LINESTRING}
\alias{gibble}
\alias{gibble.MULTILINESTRING}
\alias{gibble}
\alias{gibble.POLYGON}
\alias{gibble}
\alias{gibble.MULTIPOLYGON}
\alias{gibble}
\alias{gibble.sfc}
\alias{gibble}
\alias{gibble.sf}
\title{gibble}
\usage{
gibble(x, ...)

\method{gibble}{default}(x, ...)

\method{gibble}{POINT}(x, ...)

\method{gibble}{MULTIPOINT}(x, ...)

\method{gibble}{LINESTRING}(x, ...)

\method{gibble}{MULTILINESTRING}(x, ...)

\method{gibble}{POLYGON}(x, ...)

\method{gibble}{MULTIPOLYGON}(x, ...)

\method{gibble}{sfc}(x, ...)

\method{gibble}{sf}(x, ...)
}
\arguments{
\item{x}{geometry model}

\item{...}{arguments reserved for methods, none currently}
}
\value{
data frame summarizing the geometry map, see `Details`
}
\description{
Path-based geometry decomposition to data frame.

Geometry map in a data frame. A generic function to summarize the available paths within a
complex type. Methods are provided for the `sf` classes. There is an internal version that
is designed to work fast on sets of structures, delaying conversion to list or data frame as late as possible.
}
\details{
A gibble is map of the geometry structure of an object. Each row of the gibble represents a
component element of the object's geometry and records the number of rows (i.e. how many coordinates) and
the number of columns (i.e. the number of axes in the geometric space). Another column `type` records the
class of the object, and `object` records an identifier for the entire object. A special case column
`subobject` identifies the component POLYGON part within a MULTIPOLYGON.
I.e. `subobject` is 1 unless that part is a hole. All identifiers are sequential within implicit groups.

A gibble is not so helpful on its own, it is designed to work within other workflows such as
updating the coordinates of model object in a different context or translating between superfically
different formats.

A `gibble` is a distillation of the `sc_path` encoding of package `silicate`. That in turn came
from the  `map_table` and `sptable` decompositions of spbabel, which improve on the
single-table `fortify` model used in `ggplot2`.
}
\examples{
gibble(minimal_mesh)
}
