% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibble-package.R
\docType{data}
\name{hsh_sf}
\alias{hsh_sf}
\title{hsh, home sweet home}
\description{
A testing data set, a list of poly, line, points. The relationships
between the objects cover many interesting cases. There are multi-island
objects with holes and other objects that fall inside those holes. All paths
share vertices with a neighbour. There is topological intersection
along a diagonal line where two objects meet and are joined by
shared vertices that were inserted deliberately.
}
\details{
\code{hsh_sf} is the holey home in sf format.
}
