% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibble-package.R
\docType{data}
\name{minimal_mesh}
\alias{minimal_mesh}
\alias{minimal_mesh_sp}
\alias{minimal_mesh_sp_line}
\title{minimal mesh}
\description{
A testing data set, \code{minimal_mesh} composed of two MULTIPOLYGONs sharing one edge.
Also available are \code{minimal_mesh_sp} and \code{minimal_mesh_sp_line} in
\code{SpatialPolygonsDataFrame`` and }SpatialLinesDataFrame`` form, respectively.
}
