% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibble-package.R
\docType{data}
\name{mpoly}
\alias{mpoly}
\title{mpoly}
\description{
A testing data set, composed of two MULTIPOLYGONs, one is the manually constructed
example from sf (and silicate::sfzoo$multipolygon), the second is South Australia, with
many lake-islands in the main polygon, and one artifical hole in the Kangaroo Island polygon.
}
