% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btf.R
\name{topoedit_sf}
\alias{topoedit_sf}
\title{function to illustrate topological editing of simple features}
\usage{
topoedit_sf(x, map = NULL)
}
\arguments{
\item{x}{simple features object, supported by scsf/sc::PATH}

\item{map}{mapview object}
}
\description{
Note that this uses the sc and scsf packages which have a much
grander goal than de-duplicating vertices and re-composing simple
features. The logic used can be trivially isolated into a small set
of functions for use in mapedit, or perhaps a separate package.
}
\details{
Limited to pure edit of the points with no live redraw.
}
\examples{
x <- minimal_mesh
}
