% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibble.R
\name{gibble}
\alias{gibble}
\alias{gibble}
\alias{gibble.default}
\alias{gibble}
\alias{gibble.POINT}
\alias{gibble}
\alias{gibble.MULTIPOINT}
\alias{gibble}
\alias{gibble.LINESTRING}
\alias{gibble}
\alias{gibble.MULTILINESTRING}
\alias{gibble}
\alias{gibble.POLYGON}
\alias{gibble}
\alias{gibble.MULTIPOLYGON}
\alias{gibble}
\alias{gibble.sfc}
\alias{gibble}
\alias{gibble.sf}
\title{Gibble}
\usage{
gibble(x, ...)

\method{gibble}{default}(x)

\method{gibble}{POINT}(x, ...)

\method{gibble}{MULTIPOINT}(x, ...)

\method{gibble}{LINESTRING}(x, ...)

\method{gibble}{MULTILINESTRING}(x, ...)

\method{gibble}{POLYGON}(x, ...)

\method{gibble}{MULTIPOLYGON}(x, ...)

\method{gibble}{sfc}(x, ...)

\method{gibble}{sf}(x, ...)
}
\arguments{
\item{x}{geometry model}

\item{...}{arguments reserved for methods, none currently}
}
\value{
data frame summarizing the geometry map, see `Details`
}
\description{
Geometry map in a data frame
}
\details{
A gibble is map of the geometry structure of an object. Each row of the gibble represents a
component element of the object's geometry and records the number of rows (i.e. how many coordinates) and
the number of columns (i.e. the number of axes in the geometric space). Another column `type` records the
class of the object, and `object` records an identifier for the entire object. A special case column
`parent` identifies the parent part within a MULTIPOLYGON. I.e. `part` is 1 unless that part is a hole.

A gibble is not so helpful on its own, it is designed to work within a particular workflow such as
updating the coordinates of model object in a different context.

A `gibble` is an improvement on the `sc_path` encoding of package `sc`, which is an improvement on the `map_table`
model of `spbabel` which is an improvement on the `sptable` model of spbabel which was an improvment on the `fortify` model
in `ggplot2`.
}
\examples{
gibble(minimal_mesh)
}
