% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibble.R
\name{ibble}
\alias{ibble}
\alias{ibble}
\alias{ibble.default}
\alias{ibble}
\alias{ibble.POINT}
\alias{ibble}
\alias{ibble.MULTIPOINT}
\alias{ibble}
\alias{ibble.LINESTRING}
\alias{ibble}
\alias{ibble.MULTILINESTRING}
\alias{ibble}
\alias{ibble.POLYGON}
\alias{ibble}
\alias{ibble.MULTIPOLYGON}
\alias{ibble}
\alias{ibble.sfc}
\alias{ibble}
\alias{ibble.sf}
\title{ibble}
\usage{
ibble(x, ...)

\method{ibble}{default}(x, ...)

\method{ibble}{POINT}(x, ...)

\method{ibble}{MULTIPOINT}(x, ...)

\method{ibble}{LINESTRING}(x, ...)

\method{ibble}{MULTILINESTRING}(x, ...)

\method{ibble}{POLYGON}(x, ...)

\method{ibble}{MULTIPOLYGON}(x, ...)

\method{ibble}{sfc}(x, ...)

\method{ibble}{sf}(x, ...)
}
\arguments{
\item{x}{geometry model}

\item{...}{arguments reserved for methods, none currently}
}
\value{
matrix summarizing the geometry map, see `Details`
}
\description{
ibble is the matrix version of gibble, implemented low down for speed
}
